+++ src/app/terra-components.module.ts
@@ -47,14 +47,7 @@ import { TerraAlertComponent } from './alert/terra-alert.component';
 import { TerraAlertPanelComponent } from './alert/terra-alert-panel.component';
 import { TerraDynamicModuleLoaderComponent } from './dynamic-module-loader/terra-dynamic-module-loader.component';
 import { TerraSimpleTableComponent } from './table/simple/terra-simple-table.component';
-import { ClipboardModule } from 'angular2-clipboard/src/clipboard.module';
 import { COMPILER_PROVIDERS } from '@angular/compiler';
-import {
-    LocaleService,
-    LocalizationService,
-    LocaleModule,
-    LocalizationModule
-} from 'angular2localization';
 import { TerraTileBoxComponent } from './tile/box/terra-tile-box.component';
 import { TerraTileBoxPanelComponent } from './tile/panel/terra-tile-box-panel.component';
 import { TerraSuggestionBoxComponent } from './forms/suggestion-box/terra-suggestion-box.component';
@@ -64,7 +57,9 @@ import { TerraTextAreaInputComponent } from './forms/input/text-area-input/terra
 import { TerraLoadingSpinnerComponent } from './loading-spinner/terra-loading-spinner.component';
 import { TerraLoadingSpinnerService } from './loading-spinner/service/terra-loading-spinner.service';
 import { TerraCardComponent } from './card/terra-card.component';
+import { TranslationModule } from 'angular-l10n';
 import { TerraUrlParamsDecorator } from './service/data/terra-url-params-decorator.service';
+import { TerraToggleComponent } from './toggle/terra-toggle.component';
 export { TerraAlertPanelComponent } from './alert/terra-alert-panel.component';
 export { TerraAlertComponent } from './alert/terra-alert.component';
 export { TerraButtonInterface } from './button/data/terra-button.interface';
@@ -165,7 +160,8 @@ export { TerraCardComponent } from './card/terra-card.component';
                   TerraSuggestionBoxComponent,
                   TerraDatePickerComponent,
                   TerraTextAreaInputComponent,
-                  TerraCardComponent
+                  TerraCardComponent,
+                  TerraToggleComponent
               ],
               entryComponents: [
                   TerraTextInputComponent,
@@ -199,7 +195,8 @@ export { TerraCardComponent } from './card/terra-card.component';
                   TerraSuggestionBoxComponent,
                   TerraDatePickerComponent,
                   TerraTextAreaInputComponent,
-                  TerraCardComponent
+                  TerraCardComponent,
+                  TerraToggleComponent
               ],
               exports:         [
                   TerraAlertPanelComponent,
@@ -236,7 +233,8 @@ export { TerraCardComponent } from './card/terra-card.component';
                   TerraSuggestionBoxComponent,
                   TerraDatePickerComponent,
                   TerraTextAreaInputComponent,
-                  TerraCardComponent
+                  TerraCardComponent,
+                  TerraToggleComponent
               ],
               imports:         [
                   BrowserModule,
@@ -247,15 +245,11 @@ export { TerraCardComponent } from './card/terra-card.component';
                   TooltipModule.forRoot(),
                   AlertModule.forRoot(),
                   ButtonsModule.forRoot(),
-                  ClipboardModule,
-                  LocaleModule,
-                  LocalizationModule,
+                  TranslationModule.forRoot(),
                   MyDatePickerModule
               ],
               providers:       [
                   COMPILER_PROVIDERS,
-                  LocaleService,
-                  LocalizationService
               ],
               bootstrap:       [
                   //TerraComponentsComponent